+++ src/Symfony/Component/HttpFoundation/Session/Storage/Handler/MigratingSessionHandler.php
@@ -20,7 +20,7 @@ namespace Symfony\Component\HttpFoundation\Session\Storage\Handler;
  * @author Ross Motley <ross.motley@amara.com>
  * @author Oliver Radwell <oliver.radwell@amara.com>
  */
-class MigratingSessionHandler implements \SessionHandlerInterface, \SessionUpdateTimestampHandlerInterface
+class MigratingSessionHandler implements \SessionHandlerInterface, \SessionUpdateTimestampHandlerInterface, ClearableSessionHandlerInterface
 {
     private \SessionHandlerInterface&\SessionUpdateTimestampHandlerInterface $currentHandler;
     private \SessionHandlerInterface&\SessionUpdateTimestampHandlerInterface $writeOnlyHandler;
@@ -96,5 +96,18 @@ class MigratingSessionHandler implements \SessionHandlerInterface, \SessionUpdat
         $this->writeOnlyHandler->updateTimestamp($sessionId, $sessionData);
 
         return $result;
+    }
+
+    public function clear(): void
+    {
+        if ($this->currentHandler instanceof ClearableSessionHandlerInterface) {
+            $this->currentHandler->clear();
+        } else {
+            throw new \LogicException(\sprintf('The session handler "%s" does not support clearing all sessions.', get_debug_type($this->currentHandler)));
+        }
+
+        if ($this->writeOnlyHandler instanceof ClearableSessionHandlerInterface) {
+            $this->writeOnlyHandler->clear();
+        }
     }
 }